+++ history.md
@@ -27,6 +27,7 @@
 - 添加createCustomEdit函数创建自定义组件，添加CustomLoadEdit组件
 - 全局添加DefaultEdit的泛型M
 - 优化createCustomEdit的返回值，修正类型报错
+- SearchData添加loaded/loadFail生命周期，对应字典构建完成和失败的生命周期
 
 ### 4.8.2/3
 - 添加TrackData轨迹数据结构